chatbord/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
google-generativeai==0.3.2
python-multipart==0.0.9
pdfplumber==0.10.2
